build.sh: add variant for Linux build
This is a copy of the build.bat script with minor adaptations to build
PBSync under Linux (and possibly other POSIX systems that may have a
standard Bourne shell).

The requirements.txt file needs a bit of tweaking. pywin32 isn't needed
and pyinstaller can be added.

Updated README.md to reflect changes.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>