pbpy/pbtools.py: join the command list in the run() helpers
On Linux, running a git command (passed as a list) when shell=True, doesn't
work.
This is on a Ubuntu 20.10.
Reading the Python docs [1], it mentions that when shell=True, the command
argument should be passed as a string, and not as a sequence.

This seems more of a recommendation, than an enforcement, but with
Python 3.8 on Linux, the output is something along the lines of
"unrecognized git command"

Another idea would be to remove the 'shell=True', since most args are
passed as a list. But, the simpler workaround is to join the commands into
a string if 'shell=True'.

[1] https://docs.python.org/3/library/subprocess.html#subprocess.run
```
The shell argument (which defaults to False) specifies whether to use the shell as the program to execute.
If shell is True, it is recommended to pass args as a string rather than as a sequence.
```

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>