pbpy: add a run_with_stdin() wrapper for calling with input from stdin
This is similar to commit ca91f0a ("pbpy/pbtools.py: join the command list
in the run() helpers").

Because the 'shell=True' argument is set the commands don't work on Linux
if passed as a Python list.

To address this, we add a run_with_stdin() wrapper that handles the
conversion in the same way, and takes a parameter that is the input from
stdin.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>